feat: change appearance quickly
Due to some refactoring, now it is possible to modify the appearance of
the output on-the-fly. After changing

  g:lightline#gitdiff#indicator_added
  g:lightline#gitdiff#indicator_deleted
  g:lightline#gitdiff#separator

to your liking, simply `:w` or re-enter the buffer. The indicators will
be updated automagically i.e., there is no need to restart vim / reload
the plugin.

Kudos to /u/olminator [1] for this suggestion.

[1]: https://www.reddit.com/r/vim/comments/b7ddpd/lightlinegitdiff_show_git_diff_numstat_in_your/ejr4mea/