feat: implement calculation of modified lines
This uses `git diff --word-diff=porcelain` which is a diff per word and
not per line. This allows to calculate how many lines were added,
deleted and modified.

I split the calculation in multiple functions that parse and group the
porcelain and do the final calculation. To improve the algorithm, this
plugin definitely need a testing framework such as junegun/vader.vim.